+++ src/common/paths.ts
@@ -42,3 +42,6 @@ export const COMP_SITE_OUTDIR = `${BUILD_OUTPUT_ROOT_DIR}/comp-site` as const
 // -- Metadata file
 export const META_DATA_FILE_NAME = 'metadata.json' as const
 export const META_DATA_FILE = `${BUILD_OUTPUT_ROOT_DIR}/${META_DATA_FILE_NAME}` as const
+
+// -- Playwright testing
+export const JSON_REPORTER_FILE = './.exh/playwright/results.json'
